+++ CHANGELOG.md
@@ -2,7 +2,11 @@ Formula uses [semantic versioning](https://semver.org/).
 
 Formula is currently pre-v1. It's relatively stable, but I expect there will be some breaking changes before v1.
 
-# Unreleased
+# 0.0.9
+
+- feat: useFieldData accepts null-ish
+
+# 0.0.8
 
 - BREAKING CHANGE: useBlurred renamed to useIsBlurred
 - feat: useIsChanged, `<IsChanged>`